ci: specify permissions for release workflow
The release workflow is unable to push the tags it creates.

This seems to be caused by authentication of the GH token.

By specifying the permissions explicitly, including write access
to the workflows, the issue should not be encountered.

https://github.com/camunda/camunda/issues/28522#issuecomment-2766837108
(cherry picked from commit 119c46d8e828af46a38104aca3504c6e8792b851)