feat: Read Camunda container from Maven properties
Read the default Camunda Docker image name and version from the properties file. Use a Maven placeholder to enable
overrides for running integration tests in the CI.

The new property `camunda.dockerImageVersion` replaces the property `camunda.version` to set only the version of the
Camunda image without setting the version of the Connectors image.

Rename the version utils class to use it for all properties in the file.