fix: temporarily use the role key on role creation
It is a massive pain to use the id already since not everything is refactored to be compatible with it. Lots of tests
are failing because of it. I'd rather stick with the key here for now than disable all the authorization tests we have.
Especially since there's active work being done on Identity still. Those tests are too valuable.