refactor: store user role membership in dedicated state
This adds a new state with two column families that can store generic
membership relationships.
The first use is for users and their roles because we will eventually
have "users" that are not represented as `PersistedUser`, so we need
a new place to store which roles are assigned to these users.