feat: when reconstructing deployments, cache last resource processed
To avoid rescanning every resource every time, the last resources that
was searched for is cached in-memory (not persisted) to be used as a
starting point for the next iteration.
Persisting requires much more complexity and slows down each iteration,
while caching it in memory is slower for the first item, but faster for
all the following iterations.