fix: setCommitIndex only after flushing all the entries
Set the commitIndex of the node only after successfully flushing the
entries.
The commitIndex stored in the node should not be greater than the
lastLogIndex. Without flushing first, the lastLogIndex might not
actually be persisted successfully, so the node may refuse to reset the
log to a previous position even though it hasn't actually persisted that
data.