refactor: move tasklist tenant service outside security package
The Tasklist security package is disabled for the Camunda
single app because it conflicts with similar classes from
Operate. As a workaround, a TasklistSecurityStubsConfiguration
class was created to allow Tasklist to be used in the single
Camunda application. With the recent refactoring of the
tenant resolution mechanism, i.e. tenants are now consistenly
resolved through the service layer, the Tasklist tenant
service can be used in the single Camunda app without
conflicts. However, there are other Tasklist security
services that need to remain disabled and stubbed. As
a result, we need to move the tenant service outside
of the security package.