fix: when there is a leader transition client requests were never replied
When there is a leader transition, the LeaderAppender is closed, which
in turn terminates all pending `appendFutures`.
However, in LeaderRole, the AppendListener would call commit error only
if the role was still running, which was not true when the leader was
stepping down.
Moreover, because the future was completed with an exception, the
commitIndex is null, which would throw an exception when is implicitly
converted to `long` (i.e. not boxed).

(cherry picked from commit bcf9b5fd3df8839c578e2750b14176125e3393bf)