fix: ensure all scheduled task actors are started and stopped properly
This fixes bug where the actors running scheduled tasks were not started
and stopped properly before proceeding with the overall stream processor
startup and shutdown.

In rare cases, this would lead to SIGSEGV when a scheduled task such as
`CommandRedistributor` would try to access the runtime state after it
has already been closed.

The issue was that `AsyncUtil` returned futures that would complete with
the first step, not waiting on any further steps.

Instead of `AsyncUtil`, we now use `ActorFutureCollector` and discard
the result. Technically, this comes with a slight overhead but that's
acceptable since startup and shutdown are not in the hot path.

The broken `AsyncUtil` is removed entirely, if a new implementation is
needed, it should be tested properly and be part of already existing
utility classes such as `FutureUtil`.