feat: introduce operate adapter interface for zeebe-based services
Introduce an adapter interfaces for Operate to enable
switching between using the CamundaClient for compatibility
mode, and the CamundaServices layer, for single app mode.

This commit also introduces the Client and Service implementations
of the adapter. The implementations add the CamundaClient and the
service layer Zeebe services used by Operate.