fix: skip commit in benchmark docker build
The docker build (used in the zeebe-benchmark workflow) currently fails
as follows:

> #23 155.7 [ERROR] Failed to execute goal
> io.github.git-commit-id:git-commit-id-maven-plugin:9.0.1:revision
> (resolve-git-info) on project camunda-zeebe: .git directory is not
> found! Please specify a valid [dotGitDirectory] in your project -> [Help
> 1]

This occurs because of e7d27e16efe8dae938599d21c3c772ff5739d085 which
attempts to add git commit info to the built artifact, in combination
with the Dockerfile not having access to the .git folder.

However, for a benchmark build, we could do without this git commit
info altogether.