fix: handle error for batch execution
This adds error handling to the BatchOperationExecuteProcessor, allowing
it to deal with errors produced by the followup commands.

For example, consider an error that occurs during individual process
instance migration which is part of a batch operation. The
streamprocessor rolls back the entire result batch (and transaction), so
all progress of the BatchOperationExecution:EXECUTE command is lost
anyways. What we can do, is to append a FAILED event for this execution.
To continue, we can simply append the next followup command.

To ensure that we can continue with the next execution, the key items
must be removed from state when applying the FAILED event.

This solution has two main downsides:

1. All operations in this batch execution have FAILED, even though only
   one actually encountered a problem.

2. The reason for the encountered problem is lost.

3. If this execution was the one in the batch, then the batch operation
   is still marked as COMPLETED even though the operation execution
   FAILED.