fix: bail out when processing results were written but not committed
If committing state changes fails unexpectedly, i.e. not due to a
regular `ZeebeDbException`, we can't continue processing because log
and runtime state have diverged. We bail out by stopping processing
and indicating a dead partition.

A test is added to assert this new behavior. It uses reflection to
inject a broken transaction into the `ProcessingStateMachine` because
it did not seem worth to add dedicated testing infrastructure for this
test alone.