fix: camunda exporter - always set parentProcessInstanceKey to be able to search for root processes
Without this fix it is not possible to search root processes, as it is not allowed to use null values in the search api. (In this case, we would have to search for parentProcessInstanceKey = null)
The engine uses -1 for the parentProcessInstanceKey, so with this fix root processes with -1 as parentProcessInstanceKey can be searched.