fix: don't initialize a mapping rule
This was problematic as we had sensible defaults for the id and the claim name. However, we didn't have one for the
claim value. As a result the value would become ${INITIAL_CLAIM_VALUE} which is not desirable.

In general we don't really have to initialize a default mapping rule. It provides no value for now. We can easily re-add
 it in the future if our perspective changes.