feat: allow filtering for users by role id
We need to adjust the queries so that if a roleId is set on the filter, we actually join the ROLE_MEMBER table on the
search query.

I've also fixed the other joins by adding a check on the entity type. Since ids are not unique across entities we could
have conflicts otherwise.