feat: initialize identity by simply writing commands
We used to have this big complex processor to initialize identity. This was because we needed to know the keys in
advance when adding entities. Since we now work with ids we don't have this problem anymore. Keys can be generated
whenever and we can write all the commands using the ids instead. This greatly reduces the complexity of this
initializer, and makes the individual processors responsible for any validations.

It also means we can get rid of distributions. This is also handles by the individual processors themselves.