test: fix assertions
With the new initialize simplification these assertions need to be modified. One of the test cases was removed as it's
no longer relevant. This case verified the entities were not created again upon a restart. This is exactly what happens
now, but the individual processors are responsible for dealing with this and rejecting the command. This behaviour is
already tested elsewhere.