feat: Snapshot reservation can be persisted locally in the node
A new type of snapshot reservation has been added that allows the
reservation to be persisted, in order to survive across restarts.

The reservations for each snapshot are written to a separate file in
zeebe-reservations/${snapshotId}.bin
They are not included in the snapshot, so they are not sent to other
nodes.