fix: allow filtering by child member type
Depending on the context of the search request you're either searching sarting from the child documents, or you're
searching starting from the parent documents. The member type filter is always on the child document. This means if
you're searching from the perspective of the parent, this filter needs to be transformed to a child query. However, if
you're searching from the perspective of the child, this query needs to be transformed to the base query.

It's very confusing to have both these things in a single transformer. I believe we can refactor this to make it a lot
more clearer and intuitive which filter to use. However, due to lack of time I cannot do this right now. This is the
quick fix.