test: resolve flakiness
This test was flaky as it didn't wait for role creation and deletion to be imported to secondary storage. Retrieving
data this way would be inconsistent.

I chose to fully remove the data retrieval. The goal of this test is not to verify if the role is properly deleted from
secondary storage. The goal is to verify that the deletion is authorized properly and no unauthorized rejection is
thrown. Therefore, we can just assert that our deletion request does not throw an exception. If it doesn't that means we
 were authorized to delete the role.