fix: include operation reference for incident and job services
Extend the Incident- and JobServices of the Service API
to invlude an operation reference id when resolving
incidents and updating jobs.

This commit also adjusts the REST API Job and Incident
controllers (and their tests) to include the new
parameter. The REST API spec and Java client will be
extended with a follow-up issue (camunda/camunda#32270).