fix: temporarily provide default impl for `getVariableUpdateSemantics`
Adds a default implementation of `getVariableUpdateSemantics` in `UserTaskRecordValue`
to prevent build failures in the Optimize backend module.

This workaround is needed because the `optimize/backend` module depends on the
`zeebe-protocol` artifact version `8.8.0-SNAPSHOT`, but the CI pipeline does not build
the protocol module before building Optimize.

Reverts changes in `ZeebeUserTaskDataDto` for now, and they will be reintroduced
once the updated protocol version is available in Maven Artifactory.