feat: optionally use basic authentication
This commit adds configuration options to enable basic the use of basic auth in the starter and worker. I chose to add
an enum AuthenticationMode, as I expect in the near-future we may want to add OIDC as well.

When the authentication mode is set to BASIC we provide a credentials provider to the camunda client. This
credentials provider takes a configured username/password and adds it to requests made.

By default, we stick to no authentication. If basic auth is enabled and no credentials are configured we will be using
demo/demo. This are the default credentials used for camunda run.