feat: add new SnapshotAfterMigrationTransitionStep
SnapshotAfterMigrationTransitionStep will schedule immediately a
snapshot if any migration has been run. It will keep retrying until a
snapshot has been successfully taken, which is required if the
StreamProcessor is not processing any record.
A new component `MigrationSnapshotDirector` will be created when the
partition is transitioning to an active role. This component will be
registered as a child of SnapshotDirector and will report itself as
healthy only when a snapshot has been taken.