refactor: add a clarification comment in the job batch collector
The authorization request instance in the job batch collector
is used to retrieve the resource ids that the requester
(ex. user, application, mapping) is allowed to work with.
This data is not tenant-owned, so no tenant id needs to be
provided. Furthermore, the check that the requester is
authorized to activate jobs for the specified tenant is
performed earlier, in the corresponding processor.

This commit adds a code comment to ensure that future
developers don't get confused by the missing tenant id
in the authorizationr request.