fix: ensure internal data is not checked for tenant ownership
Internal data (like Users, Roles, Groups, Tenants) is not
owned by tenants, and tenant ownership doesn't need to be
validated in the authorization check for this type of
entities.

This commit adds a flag to the AuthorizationRequest class
to specify when data is tenant-owned. The flag is then
used to skip tenant checks.