test: avoid changing event appliers
In Zeebe, an event must be replayed in the exact same way as it was
applied to the state when it was initially written. This ensures that we
can rebuild the state from all events by replaying them. Changes to the
appliers can lead to replaying the events differently when updating to a
new version. This can lead to serious bugs as the leader and followers
may be ending up with different state. It is thus vital that we find
ways to avoid such bugs.

This test tries to warn about anyone making changes to the event
appliers. It also tries to provide instructions to overcome different
scenarios. And it provides a command to easily pass the test when you're
only just adding a new applier and there was no bug to begin with.

The test is very simple: it checks the source code of the event applier
by matching it to a golden file. If the source differs from the golden
file, the test fails. If the golden file doesn't exist, the test fails.

For the NOOP appliers, the golden file must exist but must be empty.

This test is not a full safeguard, as changes in the classes used by
each applier (e.g. state classes) could also result in different replay
outcomes. The goal of these tests is to be a first layer of defense.