feat: use the new feature flag in MessageBatchExpireProcessor
I chose to add the feature flag check as straightforward. If the ff is
enabled, we try to find the message in the state and append the full
message body. Otherwise, we append an empty message record as it was
before.

(cherry picked from commit d555f0559d451eb287186228513cd6bdf8f27404)