test: fix expected amount of users
Since my refactoring we now always create users that are annotated with a user definition annotation, regardless of if
they are specified in an authenticated parameter annotation. As a result the expected count here increases by 2.

Ideally we change this to have a more targeted query on usernames, but that's too much effort to add as part of this PR.