fix: return most relevant error on authorization check failure
Ensure that the AuthorizationCheckBehavior#isAuthorized returns
the correct error if no authorizations exist. Prevously, only
the error message of the last entity type check was returned,
which was not good UX.

To correct the issue, rejections are aggregated and permission-
related errors are prioritized since they are dependent on the
tenant-related checks.