fix: adjust user task event applier golden files
In 8.5, some changes were made to the user task event appliers compared
to 8.4. That was no problem, because the appliers could not have been
used in 8.4 (the feature was in development already, but not yet
available from outside). For example, see the 8.4 docs (they don't
mention the different implementations of user task), but 8.5 does.

However, there is a tiny chance for corruption, if the new feature is
used during a rolling update and the older follower becomes leader. This
chance is tiny, so we can likely accept it.