fix: alter golden appliers for execution listeners
Execution listeners were not available in 8.4. Any events created on 8.4
can therefore not have had any Execution Listener related state. On 8.5,
they are new, and thus the event appliers could be safely adjusted.
Let's go over the changes in more detail:

Incident_RESOLVED_v1: resets the execution listener index to 0, but it
would anyways have been set to 0 (default).

Incident_RESOLVED_v2: same as v1.

ProcessInstance_ELEMENT_ACTIVATED_v1: same case Incident Resolved

Job_COMPLETED_v1: only increments the execution listener if the job has
the EXECUTION_LISTENER kind, but no jobs of that kind could've existed
previously.

Note that there is tiny chance that these changes could lead to an
issue, when the execution listener feature is used during a rolling
update. In that case, the follower must remain on an older version while
the leader is on a newer version. Then a process with an execution
listener must be deployed, started, and the execution listener job must
be created. After which the older follower must become leader. In that
case, the execution listener index may not be correctly tracked on this
older leader. The chance for this is incredibly small, and the impact
small. We can allow it.

(cherry picked from commit 7ae6d93142fc7cf3a7fdf2050d762e9af4b3e2d6)