fix: forward claims of date-like types
This is needed unless we want to filter out claims such as `exp` and
again. These are automatically mapped to `Instant` by Spring. Without
enabling the `JavaTimeModule`, we cannot serialize these claims as
msgpack.

The fix may have a large impact because it allows _all_ msgpack values
to carry date-like types.