feat: add RawTransactionalColumnFamily to copy generically CFs from one DB to another
TransactionalColumnFamily requires knowing the types of each column,
which unfortunately is an information which is not available
generically.
RawTransactionalColumnFamily exposes some basic functionalities of a CF
with only byte[] which is ideal for copying from one RocksDB instance to
another.
RocksSnapshotDBCopy is able to open a snapshot folder, iterate over all
columns with the provided scope and insert them in another instance of
RocksDB.
At the end, it will take a snapshot in another folder.