test: verify user task state updates
This test verifies that the camunda exporter can handle each of the
user task events (based on UserTaskIntent), and updates the user task
state (CREATED, COMPLETED, CANCELED) accordingly.

The test automatically fails if new intents are added to force the
author to consider whether it should update the state as well or not.

Only events are considered, as only events inform about what has
happened.

The test uses an argument captor to ensure we only verify the
TaskTemplate.STATE property, ensuring it doesn't fail when other fields
are updated.