fix: don't set task state to created on migrated
There was no test to verify when a user task state should be updated.
Now that we have one, it failed on the MIGRATED event.

While currently the only available states are CREATED, COMPLETED, and
CANCELED, it is correct to set the state to CREATED on MIGRATED. This is
because the MIGRATED event cannot occur after the task is COMPLETED or
CANCELED, so the task is always in CREATED. In other words, only a
CREATED task can be MIGRATED.

This means that the MIGRATED event, when handled, doesn't actually
change the state to CREATED as it will always already be CREATED.

However, that's about to change when we start supporting partial states
like ASSIGNING, UPDATING, COMPLETING, etc. MIGRATED may occur at any
time before COMPLETED or CANCELED, so it may occur at any time after
these new partial states. That would mean that the MIGRATED event could
set a currently UPDATING task to CREATED even though it hasn't finished
UPDATING.

The fix is simple: just don't update the state on MIGRATED at all as it
doesn't actually change the task state.