feat: handle non-creating task states
The partial events that occur after a user task's creation only have to
update the entity to the corresponding task state.

For the UserTaskIntent.CREATING event, this is a bit different.
Previously the first event that would "create" the entity was the
CREATED event. But with the CREATING event that is about to change. This
will thus require more changes. To keep things easy to digest, we'll
split that into a separate commit.