feat: zb-snapshots for bootstrap are created in bootstrap-snapshots folder
Add support for creating snapshots in a different folder
("bootstrap-snapshots") to avoid any kind of problems with regular
snapshots.
A new field has been added to the zeebe.metadata file to indicate that
the snapshot should be used only for bootstrapping. The field defaults
to "false" so it's backward compatible.