feat: create user task entity on creating event
In order to track the CREATING event as the state of a user task, we
need to create the task entity on this event already.

That means that the CREATED event only has to adjust the state.

Note that this is not fully correct yet because the CREATING event is a
partial event and thus the assignee, due date, follow up date, priority,
candidate users, and candidate groups attributes have not yet been
committed. We should thus not store them on CREATING. We'll take care of
that in the next commit.