test: use both creating and created events
This way the test continues to verify that the entity contains all
expected properties after the CREATED event, regardless of whether we've
moved the construction to the CREATING event or not.

Note that the lifecycle of a user task always has the CREATING event. So
we don't have to worry that the CREATING event doesn't occur or so.