fix: update process version of migrated user task
If the user task is migrated, the process definition version can change.
So, it must be updated in secondary storage.

This fix also helps to pass a test that started to fail, because the
state is no longer updated to CREATED. In a previous commit I explained
that MIGRATED may now also occur after events where the task has a
different state. The state should not change on MIGRATED.