fix: only commit user task data on created
While the user task must exist in the secondary state for users to
inspect the state, it should not yet contain any user task data that can
be corrected.

This is because the partial event and any corrections may only be
committed ones all listeners have completed and the user task reaches
the full state. For creation, this is the CREATED event.

While it may be useful for users to see the partial data for the user
task already during creation, this becomes unexpected for a user task
that has been fully CREATED. For example, consider assigning foo to a
CREATED user task. The user task will then become assigning to foo,
which gives a listener the chance to approve or deny this. Until it has
made its choice, foo should not be the assignee of the user task.

We're simply applying the same rule to the creating and created events.