fix: log and exit on class circularity error
The ClassCircularityError is "Thrown when the Java Virtual Machine
detects a circularity in the superclass hierarchy of a class being
loaded". When this occurs, our applications cannot recover from it, and
a shutdown is the only reasonable solution.

We've started seeing this error occur in our virtual threads usage for
request handling in the gRPC gateway.

By expanding the VirtualMachineErrorHandler (which is used by the
UncaughtExceptionHandler for the virtual thread factory of the grpc
gateway as a FatalErrorHandler), we ensure that the gateway shutsdown as
soon as this fatal error occurs. Managed clusters like those using k8s
can then automatically recover by restarting the gateway.

Of course, the underlying bug that causes this in the gRPC gateway must
also be resolved, but we can at least support an quick shutdown when we
encounter this error.

(cherry picked from commit 299cf93dae87553b4614025f66e04d993bcab86f)