refactor: add camunda profile for operate executor testing
The Operate OperationExecutor class is started through the
StartupBean class. However, this class is not created when
the TEST profile is set. As a result, the OperationExecutor
and the operations it needs to execute can't be tested in
integration tests (that set the TEST profile).

This commit adds an additional TEST_EXECUTOR profile to
allow for the Operate OperationExecutor to be enabled
in integration tests.