test: align golden distribution appliers
It seems to me that this is fine because removeRetriableDistribution
does a deleteExisting which would result in a consistency check failure
if the acknowledged distribution was replayed after an update.

Then again it surprises me that we haven't seen that occur. Wouldn't it
be a common thing to replay command distribution acknowledged events
when updating from 8.5 to 8.6?

I'm not yet sure whether there's any impact to the
addRetriableDistribution that's newly added. Likely it doesn't actually
impact anything for replayed events.

(cherry picked from commit 2558aa73992313d186686fafb02303a6fc6b4513)