fix: limit which elements can be adhoc activities
It's currently possible to activate a boundary event existing within an
ad-hoc sub-process directly. However, this leads to an error, because
the boundary event is expected to not be activated using the ACTIVATE
command. In other words, it cannot currently be activated as an ad hoc
activity.

This change adds a filter to the ad hoc sub process transformer that
ensures only specific elements are considered activatible adhoc
activities.

The implementation uses a switch expression. By not using a default, the
compiler requires all enum values to be covered. This means that if we
ever were to expand the bpmn element or event types, then were forced to
consider it here.

(cherry picked from commit b9456de81a241d98f31c437fef28c0c74bd22187)