fix: don't cache compiled JsonPaths
A compiled JsonPath is not thread-safe, so trying to cache them becomes
very tricky. Compilation itself is reasonably fast (some
micro-benchmarks showed 300ns for compiling and 100ns for evaluation) so
it's just easier to just-in-time compile JsonPath expressions.
The evaluation cache still prevents repeated compilation and evaluation
if many mapping rules use the same expression.