refactor: enforce data consistency check on async request deletion
Use `deleteExisting` instead of `deleteIfExists` in `DbAsyncRequestState#deleteRequest` to enforce
data consistency check when async request metadata is removed from state.

We currently always call this method in contexts where the metadata is guaranteed to be present.
Switching to `deleteExisting` will validate this assumption at runtime and highlight potential
incorrect usages if they are introduced in the future.