fix: consider intermediate catch events for adhoc
While I imagined that only the none intermediate catch event would be
allowed to activate in an ad-hoc sub-process, it seems that all are
actually supported.

The expected behavior of activating an intermediate timer catch event is
that the timer triggers after the specified time since the activation of
the ad-hoc activities.