test: verify PI batch activate gets rejected if parent doesn't exist
If the parent is terminated the PI batch would result in an NPE. This results in a banned process instance. Obviously
that's not desirable and we should be able to handle this case gracefully. This test reproduces this scenario. By
writing commands/events manually we can ensure we have a deterministic ordering. We make sure we have activated the MI
body. After this we write a command to terminate the MI body, as well as another command to activate the PI batch. When
processing the PI batch activate command the MI body will be terminated and we will reach our NPE.
Instead of the NPE we expect that the command gets rejected gracefully, with a proper error message.