fix: gracefully handle non-existing parent
If the parent is terminated the PI batch would result in an NPE. This results in a banned process instance. Obviously
that's not desirable and we should be able to handle this case gracefully.
By adding a simple null-check we can determine that the parent doesn't exist anymore, and simply reject this command.