feat: allow creation of admin users
Adds a new command processor which handles a new intent. The User.CREATE_ADMIN intent. This processor will validate the
command by checking the user doesn't exist, and the admin role does exist. If both is true it will write a followup
command to create the user, and a followup command to assign the admin role to this user.