feat: add configuration option for pausing command redistribution
Add a new DistributionCfg class that provides a configuration
flag for pausing command redistribution. The flag will be used
to prevent the CommandRedistributor class from scheduling a
command distribution retry cycle runner.

The final goal is to allow Camunda Platform admins to pause
the redistribution of commands when partitions in the Zeebe
cluster are overloaded and can't process any more commands.