feat: handle /setup/user POST requests
This request will create the initial admin user. I've taken the liberty to reuse a lot of what we have already for the
regular user creation flow. This means I had to add a new controller, but the service can mostly reuse the same logic
and just use the new intent.