fix: remove hardcoded default tenant access
Adjust the retrieval method of authorized tenants.
The <default> tenant is authorized by default only
when multi-tenancy is disabled.

When multi-tenancy is enabled all tenants must be
explicitly assigned to an entity. In this case, it
is possible for an entity to be assigned to zero
tenants, i.e. the AuthorizationCheckBehavior#
getAuthorizedTenantIds(command) method can return
an empty tenantIds List.

Note: when a request is made by an anonymous user,
it isn't relevant if the multi-tenancy is enabled
or not, as anonymous users are always authorized to
perform these requests since they are made by
Camunda-internal components.