fix: ensure we iterate over event triggers to find a matching one
As proven by https://github.com/camunda/camunda/issues/33952 we have a chance that the first event trigger we find is
not actually the one related to the process instance we are activating. When this happens it doesn't mean there is no
other event trigger in the state that matches our process instance.

Instead of only looking at the first one we should iterate over the state. If we end up finding none, we can activate
the none start event. Else, we should activate the start event that belongs to the event trigger.