fix: restore correct variable map formatting in operate
Ensure that variable values are correctly passed on to
Zeebe. All variables arriving in the Operate Update
variable handler have their values converted to Strings.
When the variable name and value are formatted so they
can be passed on to Zeebe, the variable value needs to
be restored. For example, a Number String should be
converted to an Int or Long.