fix: username and email are optional fields
Email are name are optional fields. We should ensure they can be null. If they are set to null we automatically fill
them with an empty string. This also means we shouldn't check if it's empty, because if someone wants to update a user
with the API response, they'd add an empty name and email.