fix: set operationRef in incident resolve command for REST ZeebeClient
Fixes an issue where, when using REST, an empty body was sent to the resolve incident endpoint
in `ResolveIncidentCommandImpl` (deprecated ZeebeClient), which caused the `operationReference`
to be omitted from the request.

Now the command correctly sets the `operationReference` on the REST request object, and
sends an `IncidentResolutionRequest` as the request body.

Also adds missing test coverage to verify the correct behavior for both gRPC and REST requests.