fix: init user task with data on creating
While the assignee is a bit special, as we have a dedicated lifecycle for it,
the other properties don't have this, and it'd be expected that they'll
be available immediately from the start of the lifecycle.

In case if candidate users/groups were cleaned by using corrections, we would need
to clear them on a `CREATED` event.
We set `null`-s for them as this was the previous behavior on user task creation.

Co-authored-by: Nico Korthout <nico.korthout@camunda.com>